working on template rebasing

+++ ersilia/hub/fetch/pack/__init__.py
@@ -35,12 +35,15 @@ class _Symlinker(ErsiliaBase):
         os.symlink(src, dst, target_is_directory=True)
 
     def _dest_lake_symlink(self):
-        return # TODO
-        src = os.path.join(self._model_path(self.model_id), DATA_FILE)
-        dst = os.path.join(self._lake_path(self.model_id), DATA_FILE)
-        self.logger.debug("Symbolic link from {0}".format(src))
-        self.logger.debug("Symbolic link to {1}".format(dst))
-        os.symlink(src, dst, target_is_directory=False)
+        src = os.path.join(self._model_path(self.model_id), H5_DATA_FILE)
+        dst = os.path.join(
+            self._lake_path(),
+            "{0}{1}.{2}".format(self.model_id, ISAURA_FILE_TAG, H5_EXTENSION),
+        )
+        if os.path.exists(src):
+            self.logger.debug("Symbolic link from {0}".format(src))
+            self.logger.debug("Symbolic link to {1}".format(dst))
+            os.symlink(src, dst, target_is_directory=False)
 
     def _symlinks(self):
         self._bentoml_bundle_symlink()
